Move duplicate code
