+++ tests/conftest.py
@@ -29,3 +29,13 @@ def nlp():
 @pytest.fixture
 def nlp_entity(nlp):
     return _make_nlp_entity(nlp)
+
+
+class MockToken:
+    def __init__(self, text: str):
+        self.text = text
+        self.is_sent_start = False
+
+
+def get_mock_tokens(texts: list[str]):
+    return [MockToken(text) for text in texts]
